Change hkrId type to Positive Big Integer

+++ infraohjelmointi_api/models/ProjectSet.py
@@ -7,7 +7,7 @@ from django.utils.translation import gettext_lazy as lazy
 class ProjectSet(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     name = models.CharField(max_length=200, blank=True, null=True)
-    hkrId = models.IntegerField(blank=True, null=True)
+    hkrId = models.PositiveBigIntegerField(blank=True, null=True)
     description = models.TextField(max_length=500, blank=True, null=True)
     responsiblePerson = models.ForeignKey(
         "Person", on_delete=models.DO_NOTHING, blank=True, null=True
