+++ infraohjelmointi_api/models/Task.py
@@ -7,7 +7,7 @@ class Task(models.Model):
 
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     projectId = models.ForeignKey("Project", on_delete=models.DO_NOTHING)
-    hkrId = models.IntegerField(blank=True, null=True)
+    hkrId = models.PositiveBigIntegerField(blank=True, null=True)
     taskType = models.CharField(max_length=50, blank=False, null=False)
     status = models.ForeignKey(
         TaskStatus, on_delete=models.DO_NOTHING, null=True, blank=True
